+++ equinox-patterns/Domain/Period.fs
@@ -18,7 +18,7 @@ module Events =
         | Added of          ItemIds
         | CarriedForward of Balance
         interface TypeShape.UnionContract.IUnionContract
-    let codec = FsCodec.NewtonsoftJson.Codec.Create<Event>()
+    let codec = Config.EventCodec.create<Event>()
 
 module Fold =