+++ equinox-shipping/Domain/Config.fs
@@ -4,6 +4,14 @@ module Shipping.Domain.Config
 let log = Serilog.Log.ForContext("isMetric", true)
 let createDecider stream = Equinox.Decider(log, stream, maxAttempts = 3)
 
+module EventCodec =
+
+    open FsCodec.SystemTextJson
+
+    let private defaultOptions = Options.Create(autoUnion = true)
+    let create<'t when 't :> TypeShape.UnionContract.IUnionContract> () =
+        Codec.Create<'t>(options = defaultOptions).ToByteArrayCodec()
+
 module Memory =
 
     let create codec initial fold store =