+++ equinox-shipping/Domain/Shipment.fs
@@ -12,8 +12,7 @@ module Events =
         | Revoked
         | Snapshotted of {| reservation : TransactionId option; association : ContainerId option |}
         interface TypeShape.UnionContract.IUnionContract
-
-    let codec = FsCodec.NewtonsoftJson.Codec.Create<Event>()
+    let codec = Config.EventCodec.create<Event>()
 
 module Fold =