+++ propulsion-tracking-consumer/SkuSummary.fs
@@ -16,7 +16,7 @@ module Events =
         | Ingested of ItemData
         | Snapshotted of ItemData[]
         interface TypeShape.UnionContract.IUnionContract
-    let codec = FsCodec.NewtonsoftJson.Codec.Create<Event>()
+    let codec = EventCodec.create<Event>()
 
 module Fold =