+++ equinox-web/Domain/Infrastructure.fs
@@ -11,3 +11,13 @@ type ClientId = Guid<clientId>
 and [<Measure>] clientId
 module ClientId =
     let toString (value : ClientId) : string = Guid.toStringN %value
+
+/// For particularly common patterns used within a given app, sometimes it can make sense to name the pattern locally
+/// There are definitely trade-offs to this - one person's great intention revealing name is another's layer of obfuscation
+[<AutoOpen>]
+module DeciderExtensions =
+
+    type Equinox.Decider<'e, 's> with
+
+         member x.TransactWithPostState(decide, mapResult) =
+            x.TransactEx((fun c -> async { let events = decide c.State in return (), events }), fun () c -> mapResult c.State)