+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The `Unreleased` section name is replaced by the expected version of next releas
 ### Changed
 
 - Target `Propulsion` v `2.12.0-rc.3` [#111](https://github.com/jet/dotnet-templates/pull/111)
+- `eqxPatterns`: Extract `ExactlyOnceIngester` [#111](https://github.com/jet/dotnet-templates/pull/110)
 
 ### Removed
 ### Fixed