+++ CHANGELOG.md
@@ -14,9 +14,9 @@ The `Unreleased` section name is replaced by the expected version of next releas
 
 ### Changed
 
-- Target `Propulsion` v `2.12.1`, `Equinox` v `3.0.6` [#111](https://github.com/jet/dotnet-templates/pull/111) [#114](https://github.com/jet/dotnet-templates/pull/114)
+- Target `Propulsion` v `2.12.2`, `Equinox` v `3.0.6` [#111](https://github.com/jet/dotnet-templates/pull/111) [#114](https://github.com/jet/dotnet-templates/pull/114)
 - `eqxPatterns`: Extract `ExactlyOnceIngester` [#110](https://github.com/jet/dotnet-templates/pull/110)
-- Target `FsCodec.SystemTextJson`.* v `2.3.1` [#112](https://github.com/jet/dotnet-templates/pull/112)
+- Target `FsCodec.SystemTextJson`.* v `2.3.2` [#112](https://github.com/jet/dotnet-templates/pull/112)
 
 ### Removed
 ### Fixed